refactor: split zustand store into domain-specific stores
Split the monolithic conversationStore into four focused stores:
- conversationStore: phase, transcript, speech state (134 lines)
- storyStore: story content, generation, saved stories (413 lines)
- narrationStore: narration state and playback control (97 lines)
- errorStore: shared error handling across domains (68 lines)

Updated all 8 component/hook consumers:
- useConversation hook: triggers story generation via storyStore
- BookReader: uses storyStore for content, narrationStore for playback state
- StoryContent, StoryScreen: use storyStore for story/generation state
- NarrationControls: uses narrationStore for playback UI state
- ConversationInterface: uses storyStore to trigger generation on test mode
- StoryGenerationSplash: uses errorStore and storyStore separately
- bookshelf detail page: sets story via storyStore

Benefits:
- Each store ~100-400 lines (down from 580+ monolith)
- Clear responsibility boundaries reduce cognitive load
- Easier to test and refactor individual domains
- Narration state separated from conversation/story logic
- Error handling now independent and reusable

All tests pass: 21 frontend tests, 135 backend tests, lint clean.
